Реализует фильтрацию точек маршрута при помощи компонентов для фильтров
