--- src/make-filter.js
@@ -1,2 +1,0 @@
-export default (filter) => `<input type="radio" id="filter-${filter}" name="filter" value="${filter}">
-            <label class="trip-filter__item" for="filter-${filter}">${filter}</label>`;
